# public/swagger/swagger.json
{
	"components": {
...
		"schemas": {
			"IResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"error": {},
					"data": {},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"error",
					"data",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT Authorization header using the Bearer scheme."
			}
		}
	},
	"info": {
		"title": "solos_apis",
		"version": "1.0.0",
		"description": "solos apis",
		"license": {
			"name": "license"
		},
		"contact": {
			"name": "BTB"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/fred/createUpdateFredNFT": {
			"post": {
				"operationId": "CreateUpdateFredNFT",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Create / Update Fred NFT, and Type = \" flexible / fixed \"",
				"tags": [
					"Fred"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"description": {
										"type": "string"
									},
									"fred_name": {
										"type": "string"
									},
									"benefit_month": {
										"type": "string"
									},
									"years": {
										"type": "string"
									},
									"amount": {
										"type": "string"
									},
									"benefit_amount": {
										"type": "string"
									},
									"purchase_limit": {
										"type": "string"
									},
									"start_month": {
										"type": "string"
									},
									"end_year": {
										"type": "string"
									},
									"type": {
										"type": "string"
									},
									"fred_nft_id": {
										"type": "string"
									},
									"tokenId": {
										"type": "string"
									},
									"contractAddress": {
										"type": "string"
									},
									"maturityDate": {
										"type": "string"
									},
									"nft_image": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/fred/nftTransfer": {
			"post": {
				"operationId": "NftTransfer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Normal NFT transfer to user",
				"tags": [
					"Fred"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"signerAddress": {
										"type": "string"
									},
									"type": {
										"type": "string"
									},
									"amount": {
										"type": "string"
									},
									"nftDetailId": {
										"type": "string"
									},
									"hashId": {
										"type": "string"
									},
									"receiverId": {
										"type": "string"
									}
								},
								"required": [
									"signerAddress"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/fred/listFredNFT": {
			"get": {
				"operationId": "ListFredNFT",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "List with Pagination Fred NFT",
				"tags": [
					"Fred"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page_number",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page_limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "filter",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"0",
								"1"
							]
						}
					}
				]
			}
		},
		"/api/fred/deleteFredNFT": {
			"delete": {
				"operationId": "DeleteFredNFT",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Delete Fred NFT",
				"tags": [
					"Fred"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/fred/purchaseNftAndFredFromUser": {
			"post": {
				"operationId": "PurchaseNftAndFredFromUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Purchase nft and fred from user",
				"tags": [
					"Fred"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"type": {
										"type": "string"
									},
									"amount": {
										"type": "string"
									},
									"nftDetailId": {
										"type": "string"
									},
									"hashId": {
										"type": "string"
									},
									"receiverId": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/fred/airdropFredNft": {
			"put": {
				"operationId": "AirdropFredNft",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Fred"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"contractAddress": {
										"type": "string"
									},
									"userId": {
										"type": "string"
									},
									"FredId": {
										"type": "string"
									}
								},
								"required": [
									"contractAddress",
									"userId",
									"FredId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/fred/airdropNft": {
			"put": {
				"operationId": "AirdropNft",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Fred"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"contractAddress": {
										"type": "string"
									},
									"userId": {
										"type": "string"
									},
									"NftId": {
										"type": "string"
									}
								},
								"required": [
									"contractAddress",
									"userId",
									"NftId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}